+++ include/jsoncons_ext/jsonpath/flatten.hpp
@@ -113,7 +113,7 @@ namespace jsonpath {
 
         if (JSONCONS_UNLIKELY(!value.is_object()))
         {
-            JSONCONS_THROW(jsonpath_error(jsonpath_errc::argument_to_unflatten_invalid));
+            JSONCONS_THROW(jsonpath_error(jsonpath_errc::invalid_argument_to_unflatten));
         }
 
         Json result;